Allow customizing logged span events (#40)
Currently, new/close span events are logged (that is, when a span is
created for the first time, usually when an `#[instrument]`ed function
is called, and when spans are closed, usually when those functions
return). It can be useful to disable those for terser logs, or to enable
more events for greater detail.

This change adds a `--trace-spans` argument to customize which span
events are logged.

Before shipping 1.0, we should clean up the option names. Maybe
`--tracing-filter` could be `--log` or something.