Remove `GhciStdin` process
This removes the process that manages `GhciStdin` and instead replaces
the handle to that process with the `GhciStdin` struct.

The motivation behind this change is to remove the need for
a `StdinEvent` type and avoid the need to spawn and monitor/restart a
process.  This also simplifies method invocations on that process, too.