Fix borrow-checking errors
The overall approach I took was to give the `Ghci` struct ownership
of the `stdout` value and have all of the `GhciStdin` methods that need
it borrow it mutably.

This makes the `GhciStdin` methods more verbose (they now need an
explicit `stdout` parameter which needs to be threaded throughout),
but the long-term goal is that this will be a net simplification when
we can get rid of the sender/receiver channels completely.