Fewer processes and `Sender`s (#49)
The main motivation for this change was to remove `Sender`-based
result-passing in favor of passing values directly through return
values. However, in order to do this I had to remove the subprocesses
that were managing `GhciStdout` and `GhciStderr`.

The basic approach I used to make this work and appease the
borrow-checker was:

- Make the `Ghci` struct own all three of `stdin`/`stdout`/`stderr`
- Any method that required more than one of them would borrow the other
ones it was missing

I think there's actually a better way to do this that doesn't require as
much parameter passing, which is to move the methods on
`GhciStdin`/`GhciStdout`/`GhciStderr` to become methods on `Ghci`, but
I'll save that for a separate refactor since this change was already
starting to get a big large. I picked this approach for my first pass
since it led to the smallest diff for ease of review.

Then once I finished that refactor I was able get rid of a lot of the
`Sender`-related logic and instead pass the results that would have
normally been passed via `Sender`s via the function's return value
instead. If you squint you can sort of think of this as analogous to
converting code from "continuation-passing style" to "direct style".