Add tracing JSON reader to test-harness (#62)
The Tracing JSON reader is an async task which will wait for `ghcid-ng`
to write tracing log events to a file, deserialize them, and make them
available to the test harness.

This is the core of the tests; `ghcid-ng` is started, we wait for a
particular log event, we make changes with the IO helpers, repeat.