Add test harness (#57)
The `GhcidNg` type copies a given directory to a tempdir and then
launches `ghcid-ng` there. It also spawns an async task to read JSON log
events from the log file.

`GhcidNg` uses thread-local storage to determine which version of `ghc`
to use when launching tests, and will error appropriately if the
thread-local storage is not set. This takes advantage of the fact that
async tests run in the `tokio` "current-thread" runtime by default,
which schedules all tasks in the test on the same thread.

The second part is a proc macro attribute exported as
`#[test_harness::test]`, which rewrites tests so that

1. The tests are async functions which run under the default
`#[tokio::test]` current-thread runtime.
2. Tracing is set up in the tests so that log messages from the
`test-harness` library are visible in the test output.
3. The appropriate thread-local variables are set for each test (this
includes the current GHC version).
4. One test is generated for each GHC version.
5. If tests fail, the relevant `ghcid-ng` logs are saved to a directory
under `target/` and the path is printed at the end of the tests.

Here's a sample test from #44 using this test harness:

```rust
/// Test that `ghcid-ng` can start up and then reload on changes.
#[test]
async fn can_reload() {
    let mut session = GhcidNg::new("tests/data/simple")
        .await
        .expect("ghcid-ng starts");
    session
        .wait_until_ready()
        .await
        .expect("ghcid-ng loads ghci");
    fs::append(session.path("src/MyLib.hs"), "\n\nhello = 1\n")
        .await
        .unwrap();
    session
        .wait_until_reload()
        .await
        .expect("ghcid-ng reloads on changes");
    session
        .get_log(
            Matcher::span_close()
                .in_module("ghcid_ng::ghci")
                .in_spans(["on_action", "reload"]),
        )
        .await
        .expect("ghcid-ng finishes reloading");
}
```

---------

Co-authored-by: Gabriella Gonzalez <gabriella@mercury.com>