Make `--errors` output compatible with `ghcid` (#78)
The biggest user-facing change here is to replace the `Ok, 123 modules
loaded.` "compilation summary" headline with `All good (123 modules)`;
[the `ghcid` editor plugins are hard-coded to recognize this
string.](https://github.com/ndmitchell/ghcid/blob/e2852979aa644c8fed92d46ab529d2c6c1c62b59/plugins/vscode/src/extension.ts#L79)

Previously, I had a rather clever system for the error log which would
direct anything on stderr to the error log, divided up by what it was
doing; a separate section for test output and compiler output. This is
of dubious utility -- anyone who wants to see the test output can just
look at the console, so the `--errors` output should be for tooling
only. In the future, it might be nice to experiment with getting test
failures in there too -- maybe we could reformat them to look like
compiler errors?

This change wouldn't have been so big, but changing the way the error
log is written to be purely a function of the parsed diagnostics meant
it was easy to factor out into a new type, leaving the `GhciStderr` task
stripped down and greatly simplified.

Also, I needed to write `Display` impls for the GHC diagnostics, which
necessitated a new `GhcDiagnostic` type and lots of tests.