Add eval support (#86)
Adds support for evaluating commands in source files on reload.

This looks for line commands starting with `-- $>` or multiline commands
delimited by `{- $>` and `<$ -}`.

This also removes the `failed_modules` mechanism from `Ghci`, which was
populated by extracting filenames from error messages, and replaces it
with a `targets` `ModuleSet` populated by `:show targets`. This should
be more consistent.

Finally, paths will be printed as relative paths when possible.