Fix "modules defined in multiple files" bug when modules are renamed (#125)
This fixes a "modules defined in multiple files" bug when a module is
renamed. The rename is interpreted as one module being deleted (which
triggers a restart) and one module being added (which triggers an
`:add`).

The bug is that it would restart `ghci`, thereby loading the new module,
and then attempt to `:add` the new module anyways, which would fail.