Remove sync feature (#128)
The sync feature works by instructing `ghci` to `putStrLn
"###~GHCID-NG-SYNC-123~###"`, where `123` is a "sync ID". Then it waits
for that marker to be printed on stdout, so that stdin and stdout are
"in sync" — all the input before the sync instruction has been
processed.

This is copied from ghcid and doesn't appear to be needed here, as we
are careful with how we look for prompts.