Fix error when reading from stdout splits UTF-8 codepoint (#144)
When we read from stdout, we get some number of bytes. In some cases,
when `ghci` is outputting non-ASCII characters, this can lead to a
single UTF-8 codepoint (represented as multiple bytes) being split
across two reads. (Or potentially more, in perverse cases?)

Fortunately, the `std` UTF-8 decoder makes this information available
and allows us to distinguish between truncated continuation sequences
and truly invalid data. For truncated continuation sequences, we decode
as much as we can (fortunately the `std` UTF-8 decoder makes this
information available as well) and save the last few bytes (we know
there are only 1-3 of them) for the next time we read data.