Graceful shutdown (#97)
This PR switches `ghcid-ng` from the `watchexec` crate to using the
`notify` crate directly. (This is what `watchexec` uses under the hood.)
`watchexec` is very nice, but it's largely designed around the needs of
the `watchexec(1)` binary (namely running commands when files change).
Using `notify` makes the code much nicer.

One of the things that `watchexec` provides that `notify` doesn't is
shutting down when Ctrl-C is pressed. Most of the rest of the changes
are about implementing this graceful shutdown procedure.

See [the “Graceful Shutdown” section in the Tokio
book](https://tokio.rs/tokio/topics/shutdown) for more details.

Originally I implemented this with [the `tokio-graceful-shutdown`
crate](https://docs.rs/tokio-graceful-shutdown/latest/tokio_graceful_shutdown/),
but it turned out to be missing some features I wanted. (In particular,
performing a graceful shutdown when Ctrl-C is pressed and then
performing a hard shutdown if Ctrl-C is pressed again.) Fortunately,
implementing the subset of features we need wasn't too hard.