Parse eval commands when they end with EOF instead of a newline (#153)
With an eval command at the end of the file and no newline at the end of
the file, parsing fails.

This happens because [VS Code is a bad text
editor](https://stackoverflow.com/questions/44704968/visual-studio-code-insert-newline-at-the-end-of-files)
that doesn't respect
[POSIX](https://stackoverflow.com/questions/729692/why-should-text-files-end-with-a-newline),
which defines a line as:

> **3.206 Line**
> 
> A sequence of zero or more non- \<newline> characters plus a
terminating \<newline> character.

No terminating newline? Not a line.

Anyways, we have to work around this because people aren't exactly going
to stop using VS Code.