Make GHCi output destination configurable (#175)
Context:

I'm working on adding a TUI mode, and I'd like GHCi's output to go to an
in-memory buffer instead of to `stdout` and `stderr`, so I can render it
as I see fit.

This change will allow me to drop in a pair of
`tokio::io::DuplexStream`s in place of `tokio::io::{Stdout,Stderr}`.

Worth mentioning that this does _not_ affect output from `tracing`. It
only controls the `ghci` process' output.