Check all events in `logged_in_checkpoint` (#218)
In the test harness, we use methods like `assert_logged_in_checkpoint`
and `assert_logged_or_wait` to find events in a checkpoint. These
methods error when no matching event is found and when the matcher
errors.

Matcher errors are important -- the `NegativeMatcher` will error if its
contained matcher matches an event, so we need to:

1. Check every event in the checkpoint when using these methods, even if
we find a matching event early.
2. Have a separate method that only errors if the matcher errors (to
distinguish between "matching failed" and "no matching event was found")
— this lets `assert_logged_or_wait` distinguish between a matcher error
(which should be propagated) and a missing event (so that it can wait
for an event to be logged).