`Module` -> `CompilingModule`
This is the result of parsing a GHC message like

    [1 of 3] Compiling Foo ( Foo.hs, Foo.o, interpreted )

It is specifically for when a module is being compiled. In the future,
we may parse out output names, the "1 of 3" bit, etc.

I'm thinking about adding a more complete `Module` type to go in the
`ModuleSet` and I don't want name conflicts.