Don't restart for removed or renamed modules
Ghcid crashes when modules are removed/renamed because all it can do to
the underlying GHCi session is `:reload`, and a GHCi bug [1] means that
the GHCi session will crash when this happens.

Because ghciwatch tracks which files/modules are loaded in the session,
it can accurately `:unadd` files from the session when they're removed
on disk, without requiring the session be restared.

This should speed up development.

[1]: https://gitlab.haskell.org/ghc/ghc/-/issues/11596