Add hidden debug info display
Separate keys for enabling and disabling, because holding the key down
for too long counts as multiple inputs, and causes the debug info to
flicker.

Fix is straightforward - either ignore `KeyEventKind::Repeat`, or record
a past `KeyEventKind::Press` and wait for a `KeyEventKind::Release`
before toggling again - but that's too much work right now for this
hacky debug thing.